Refactor LLaDA components and remove unused files
- Deleted chat.py and get_log_likelihood.py as they are no longer needed in the current architecture.
- Updated generate.py to utilize LLaDAModelLM instead of AutoModel for improved model handling.
- Introduced p2_plus_sampling function in p2.py to enhance sampling capabilities with fixed decoding order.
- Added new scripts for evaluating LLaDA models on GSM8K and Humaneval tasks, improving usability and testing workflows.
- Introduced kv-cache-mdm.md to outline the implementation plan for a Key-Value cache mechanism in the Masked Diffusion Model.

+++ src/path_planning/score_function.py
@@ -44,7 +44,7 @@ def random_score(logits: torch.Tensor, x0: torch.Tensor) -> torch.Tensor:
     Returns:
         Tensor of shape (batch_size, seq_len) - Random log scores for each position
     """
-    return torch.rand_like(x0).log()
+    return torch.rand_like(x0.float()).log()
 
 
 def diff_top2(logits: torch.Tensor, x0: torch.Tensor) -> torch.Tensor:
